working on help
almost there.
we now have a nice StealAnswerer struct such that we end with a single
sequential iterator on all not stolen micro blocks.
it is then flattened and the main sequential thread is now just a
standard fold.

now we need to duplicate this trick for helper threads where we can be
both stolen or retrieved.
then we can remove the now obsolete Taker struct from the api.