divide_on_right, divide_on_left_at
been thinking about par_bridge.
we can have a nice adaptive iterator where we never store any item in a
vec if not stolen. we just need to keep the sequential iterator inside
the parallel iterator.
when stolen we extract half the items from the iterator, keep them on
the left and give the iterator on the right.

it's nice but there is a problem when using macro blocks.
in this case the sequential iterator needs to move between blocks.
(at end of first block, we move it on second block).
in order to do that however we need to change the second's block
par_iter "at distance" (same pb as successors iterator but on another
level).

-> for macro blocks we need a borrow (and a streaming iterator).

some other adaptors can also benefit from having a special division
method for the macro blocks (cap, level).

we already have these and have just been renamed to tell explicitly
when dividing on the left side (sequential block iterator) or right side (//ism).