moved files, power is back
we need the power in the ItemProducer because I could not find a way
to propagate the information that a type is indexed towards it's
borrowed types.
I tried to add a constraint in IndexedParallelIterator like
for<'e> <Self as Borrowed<'e>>::ParIter : Indexed
but this seems to create a cycle since I cannot even get the ParRange to
compile.
my guess is that if I want to have an indexed ParRange then I first need
to have an Indexed ParRange.

so, I went the easy way and stored a power for each iterator as in
new_api branch.