eitheriter, chain, bugfix in logs
chain is pretty nice. i like it.

what we do is use the macro blocks to always separate on the boundary
between the two iterators.
this way all borrowed iterators and all sequential iterators are purely of one type.
might even be faster that the one in new_api.

it was designed for speed though but to solve the following pb:
since we cannot divide at an index how to divide a chain ???

by doing it in the current way the exact division is done by par_borrow.

this required patching the reduce method on parallel iterators to use
blocks